+++ app/app.js
@@ -7,6 +7,8 @@ Main app file for modbus-client
 
 const settings      = require('./settings');
 const express       = require('express');
+const bodyParser    = require('body-parser');
+const router        = express.Router();
 const path          = require('path');
 const morgan        = require('morgan');
 const compression   = require('compression');
@@ -24,15 +26,26 @@ morgan.token('prefix', function getId() {
 });
 app.use(morgan(':prefix :method :status :remote-addr :url'));
 
+app.use(bodyParser.urlencoded({extended: false}));
+app.use(bodyParser.json());
 app.use(compression());
 app.use(express.static(path.join(__dirname, 'public')));
 app.use('/', require('./routes/index'));
 console.log(settings);
 
+router.post('/:deviceId/:elementId', function (req, res) {
+  modbusDevices.edit(req.params.deviceId, req.params.elementId, req.body, err => {
+    res.status(200).send('ok');
+  });
+});
+
 modbusDevices.init(settings, socket.init(server), err => {
   if (err) {
     logger.error(err);
   }
+
+  app.use('/edit', router);
+
   server.listen(settings.port, 'localhost', () => {
     logger.info(`Server started on Port ${settings.port}`)
   });