+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #Release Notes modbus-client
 
+ * Writing data to a modbus server
+
 V0.9.0, 22.02.2018: Beta Release
  * First release with basic functionality
  * Supported registers: